+++ UIComponents.Generators/Configuration/UICConfigOptions.cs
@@ -25,7 +25,13 @@ public partial class UicConfigOptions
     #endregion
 
 
-
+    #region Events
+    public event EventHandler OnVersionChanged;
+    public async void TriggerOnVersionChanged()
+    {
+        OnVersionChanged?.Invoke(null, EventArgs.Empty);
+    }
+    #endregion
 
     #region Properties
 
@@ -62,6 +68,8 @@ public partial class UicConfigOptions
     public bool AddChangeLog { get; set; }
 
     public bool AddFileExplorerImgs { get; set; }
+
+    public bool AddTranslationFile { get; set; }
 
     public bool CheckLanguageServiceType { get; set; } = true;
     public bool CheckPermissionServiceType { get; set; } = true;