+++ UIComponents.Generators/Generators/Property/Inputs/UICGeneratorInputSelectList.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.Logging;
 using System.Collections;
 using UIComponents.Abstractions.Extensions;
 using UIComponents.Abstractions.Interfaces.ValidationRules;
+using UIComponents.Abstractions.Varia;
 using UIComponents.Generators.Helpers;
 using UIComponents.Models.Extensions;
 
@@ -36,14 +37,14 @@ public class UICGeneratorInputSelectList : UICGeneratorProperty
             input.Value = ((int)args.PropertyValue).ToString();
 
         input.ValidationRequired = await _validationService.ValidatePropertyRequired(args.PropertyInfo, args.ClassObject);
-        input.SelectListItems = (await args.Configuration.GetSelectListItems(args, input))?.ToUIC()?? new();
+        input.SelectListItems = (await args.Configuration.GetSelectListItems(args, input))?? new();
 
-        if((args.Options.SelectlistAddEmptyItem ||!input.ValidationRequired) && input.SelectListItems.Where(x => string.IsNullOrEmpty(x.Value?.ToString()??null)).Any())
+        if((args.Options.SelectlistAddEmptyItem ||!input.ValidationRequired) && !input.SelectListItems.Where(x => string.IsNullOrEmpty(x.Value?.ToString()??null)).Any())
             input.SelectListItems.Insert(0, new());
 
 
         if(input.Placeholder == null)
-            input.Placeholder = new Translatable("Select.PlaceHolder", "Select a {0}", TranslationDefaults.TranslateType(args.PropertyType));
+            input.Placeholder = TranslatableSaver.Save("Select.PlaceHolder", "Select a {0}", TranslationDefaults.TranslateType(args.PropertyType));
         
         if (args.PropertyType.IsEnum || args.CallCollection.Caller is not UICInputGroup)
             showButtonAdd = false;
@@ -60,7 +61,7 @@ public class UICGeneratorInputSelectList : UICGeneratorProperty
                     inputGroup.AppendInput.Add(new UICButton()
                     {
                         AppendButtonIcon = IconDefaults.Add,
-                        Tooltip = new("Button.CreateOfType.Tooltip", "Create a new {0}", TranslationDefaults.TranslateType(propertyType)),
+                        Tooltip = TranslatableSaver.Save("Button.CreateOfType.Tooltip", "Create a new {0}", TranslationDefaults.TranslateType(propertyType)),
                         OnClick = new UICActionGetPost(UICActionGetPost.ActionTypeEnum.Get, propertyType.Name, "create")
                         {
                             OnSuccess = new UICActionOpenResultAsModal()